Tested working on the Teensy LC.
Added and tested code to read buttons off analog input channel.

Moved vibration motor pins as these were conflicting with serial channels 2 & 3 pins. Originally intended for them to be on A8 and A9, but misunderstood that the AnalogWrite function takes a pin number and not the analog channel number (like AnalogRead).